First large improvement in classification performance.
1. Used information from both convolution stages (after pooling) as
suggested in the paper "Traffic Sign Recognition with Multi-Scale
Convolutional Networks" by Sermanet and LeCun.

2. Increased the size of the classifier on top of the convolution
stages with the intention of increasing the 'memorization' capability of
the network (previously I noticed the performance was capped at about
5%, and thought that maybe this was because each training batch was
causing the network to learn different parts of the data set and maybe
overwriting what was previously learned).

This was done simultaneously so I'm not sure at this moment which change
caused the significant improvement.

Note that I trained for 50 epochs with a batch size of 256. After only
10 epochs, the accuracy on the validation set was ~26%. Previously I had
only regularly managed to achieve 5-10% with this number of training
epochs.